implements unix socket support.
encountered multiple issues: improper serialization of strings for the
socket address using unsafe. fixed by converting the data structure to
use a fixed byte array. though the size is likely incorrect. made a
similar adjustment for ip6 zones.

finally we remap the socket path to the path of the host.